% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expected_diff.R
\name{compute_expected_diff}
\alias{compute_expected_diff}
\title{Evaluate expected change in outcome given a shift in exposure values.}
\usage{
compute_expected_diff(fit, x1, x2, z1 = NULL, z2 = NULL)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{x1}{Numeric vector of exposure values}

\item{x2}{Numeric vector of exposure values}

\item{z1}{(Optional) Numeric vector of covariate values}

\item{z2}{(Optional) Numeric vector of covariate values}
}
\value{
A vector of posterior samples of the expected outcome given a shift from x1 to x2. (And optionally from z1 to z2.)
}
\description{
Evaluate expected change in outcome given a shift from exposures x1 to exposures x2. Optionally can take covariates z1 and z2, but these don't have an effect on the result if z1=z2.
}
