% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expected_outcome.R
\name{compute_expected_outcome}
\alias{compute_expected_outcome}
\title{Evaluate expected outcome given exposure values}
\usage{
compute_expected_outcome(fit, x_obs, z_obs)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{x_obs}{Numeric vector of exposure values}

\item{z_obs}{Numeric vector of covariate values}
}
\value{
A vector of posterior samples of the expected outcome
}
\description{
Evaluate expected outcome given exposure values
}
