% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LowFR.R
\name{fit_LowFR}
\alias{fit_LowFR}
\title{Fit a Low-Rank Longitudinal Factor Regression Model}
\usage{
fit_LowFR(
  y,
  X,
  p,
  TT,
  Z = NULL,
  k = NULL,
  output = "all",
  burnin = 1000,
  samples = 1000,
  chains = 4,
  random_seed = 1234
)
}
\arguments{
\item{y}{Numeric vector of output values.}

\item{X}{Numeric matrix of exposures.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit a Low-Rank Longitudinal Factor Regression Model
}
