% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowFRPackage-package.R
\docType{_PACKAGE}
\name{LowFRPackage-package}
\alias{LowFRPackage-package}
\alias{LowFRPackage}
\title{The 'LowFRPackage' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
