% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covariate_effects.R
\name{plot_covariate_effects}
\alias{plot_covariate_effects}
\title{Plot covariate effects for all exposures}
\usage{
plot_covariate_effects(fit, varnames, lower = 0.025, upper = 0.975)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{varnames}{q dimensional vector of names of exposures}

\item{lower}{(Optional) Lower quantile value between 0 and 1}

\item{upper}{(Optional) Upper quantile value between 0 and 1}
}
\value{
A ggplot object plotting main effect posterior means and credible intervals.
}
\description{
Plot posterior means and credible intervals for the covariate effects.
}
