% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_effects.R
\name{plot_cumulative_effects}
\alias{plot_cumulative_effects}
\title{Plot cumulative effects for all chemicals}
\usage{
plot_cumulative_effects(fit, varnames, lower = 0.025, upper = 0.975)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{varnames}{p dimensional vector of names of chemicals}

\item{lower}{(Optional) Lower quantile value between 0 and 1}

\item{upper}{(Optional) Upper quantile value between 0 and 1}
}
\value{
A ggplot object plotting cumulative effect posterior means and credible intervals.
}
\description{
Plot posterior means and credible intervals for the change in expected outcome if each chemical is shifted from -0.5 to 0.5 at all times, with all other chemicals held constant at 0.
}
