% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot effects for pairs of times}
\usage{
plot_heatmap(
  fit,
  chems,
  t1,
  t2,
  lower = 0.025,
  upper = 0.975,
  zero_out_insignificant = TRUE,
  grid_resolution = 0.05,
  grid_min = -2,
  grid_max = 2
)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{chems}{Vector of indices between 1 and p indicating which chemicals to vary}

\item{t1}{Integer between 1 and TT indiating the first time at which to vary exposure}

\item{t2}{Integer between 1 and TT indicating the second time at which to vary exposure}

\item{lower}{(Optional) Lower quantile value between 0 and 1; determines interval width for zeroing out heatmap}

\item{upper}{(Optional) Upper quantile value between 0 and 1; determines interval width for zeroing out heatmap}

\item{zero_out_insignificant}{(Optional) Indicates whether to set regions where the specified credible interval includes zero equal to zero.}

\item{grid_resolution}{(Optional) The resolution at which to make the heatplot}

\item{grid_min}{(Optional) Minimum exposure level on axes}

\item{grid_max}{(Optional) Maximum exposure level on axes}
}
\value{
A ggplot object plotting a heatmap of the expected change in the outcome as the specified chemicals vary away from zero, with fixed covariate values and all other chemicals fixed at 0.
}
\description{
Plot a heatmap of posterior mean expected shifts in y when a list of chemicals varies around zero at a pair of observation times, with all other exposures held constant at zero. Regions where the 95\% credible interval includes zero are shown as white.
}
