% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main_effects.R
\name{plot_main_effects}
\alias{plot_main_effects}
\title{Plot main effects for all exposures}
\usage{
plot_main_effects(fit, varnames, lower = 0.025, upper = 0.975)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{varnames}{p*TT dimensional vector of names of exposures}

\item{lower}{(Optional) Lower quantile value between 0 and 1}

\item{upper}{(Optional) Upper quantile value between 0 and 1}
}
\value{
A ggplot object plotting main effect posterior means and credible intervals.
}
\description{
Plot posterior means and credible intervals for the change in expected outcome if each exposure is shifted from -0.5 to 0.5, with all other exposures held constant at 0.
}
