% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_covariate_effects.R
\name{summarize_covariate_effects}
\alias{summarize_covariate_effects}
\title{Summarize covariate effects}
\usage{
summarize_covariate_effects(fit, varnames, lower = 0.025, upper = 0.975)
}
\arguments{
\item{fit}{Model fit returned by fit_lowFR function with default setting "all"}

\item{varnames}{q dimensional vector of names of covariates (i.e., the columns of Z)}

\item{lower}{(Optional) Lower quantile value between 0 and 1}

\item{upper}{(Optional) Upper quantile value between 0 and 1}
}
\value{
A dataframe with p*TT rows summarizing each main effect.
}
\description{
Compute posterior means and credible intervals for the linear covariate effects.
}
