% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_posterior.R
\name{summarize_posterior}
\alias{summarize_posterior}
\title{Summarize a vector of posterior samples in terms of the posterior mean and credible interval}
\usage{
summarize_posterior(samples, lower = 0.025, upper = 0.975)
}
\arguments{
\item{samples}{Vector of posterior samples of some quantity of interest}

\item{lower}{Lower quantile value between 0 and 1}

\item{upper}{Upper quantile value between 0 and 1}
}
\value{
A list containing the mean and lower and upper quantiles (default 2.5th and 97.5th percentile, giving an equal-tailed 95\% interval)
}
\description{
Summarize a vector of posterior samples in terms of the posterior mean and credible interval
}
