% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LowFR.R
\name{fit_LowFR}
\alias{fit_LowFR}
\title{Fit a Low-Rank Longitudinal Factor Regression Model}
\usage{
fit_LowFR(
  y,
  X,
  p,
  TT,
  Z = NULL,
  k = NULL,
  output = "all",
  preestimate_Lambda = FALSE,
  burnin = 1000,
  samples = 1000,
  chains = 4,
  random_seed = 1234
)
}
\arguments{
\item{y}{Numeric vector of outcome values}

\item{X}{Numeric matrix of exposure values. Can include NAs for missing values.}

\item{p}{Number of different exposures (e.g, chemicals) stored in X.}

\item{TT}{Number of different measurement times stored in X}

\item{Z}{(Optional) Numeric matrix of additional covariate values.}

\item{k}{(Optional) Number of latent factors to use. (Will be estimated with SVD if not provided.)}
}
\value{
A list of arrays containing posterior samples.
}
\description{
Fit a Low-Rank Longitudinal Factor Regression Model
}
